desugar_guard_b.rs hands the record-literal arg hoist to desugar_guard_c.rs, a pure text move under the 800-line ceiling
