+++ crates/almide-mir/src/lower/mod_c.rs
@@ -980,6 +980,7 @@ include!("newtype_subst.rs");
 include!("record_defaults.rs");
 include!("desugar_guard.rs");
 include!("desugar_guard_b.rs");
+include!("desugar_guard_c.rs");
 include!("cells.rs");
 include!("inline_scalar_fns.rs");
 include!("mod_p2.rs");
